requirements.txt
python-dotenv
pydantic
requests
alembic
sqlalchemy
pyodbc
cryptography
fastapi
uvicorn
google-genai
duckduckgo-search
openai
beautifulsoup4
html5lib
python-jose
bcrypt